examples/langchain_agent/requirements.txt
langchain>=0.2.14
langchain-openai>=0.1.22
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
